fix: build command was excluded from backups (#4)
build is excluded from the backups (tar) this is applicable to folders and files.
build is often used as command in my projects. The build folder contains temporary files for creating an executable.
Changed exclusion to exclude all contents in a build folder, it does not exclude build as a file/command.